add examples

+++ include/blackwidow/blackwidow.h
@@ -46,7 +46,7 @@ class BlackWidow {
     kBitOpOr,
     kBitOpXor,
     kBitOpNot,
-    kBitOpDefault 
+    kBitOpDefault
   };
 
   // Set key to hold the string value. if key
@@ -57,7 +57,7 @@ class BlackWidow {
   // the special value nil is returned
   Status Get(const Slice& key, std::string* value);
 
-  // Atomically sets key to value and returns the old value stored at key 
+  // Atomically sets key to value and returns the old value stored at key
   // Returns an error when key exists but does not hold a string value.
   Status GetSet(const Slice& key, const Slice& value, std::string* old_value);
 
@@ -83,7 +83,8 @@ class BlackWidow {
   Status Setnx(const Slice& key, const Slice& value, int32_t* ret);
 
   // Sets the given keys to their respective values.
-  // MSETNX will not perform any operation at all even if just a single key already exists.
+  // MSETNX will not perform any operation at all even
+  // if just a single key already exists.
   Status MSetnx(const std::vector<BlackWidow::KeyValue>& kvs, int32_t* ret);
 
   // Set key to hold string value if key does not exist
@@ -107,27 +108,27 @@ class BlackWidow {
   Status BitCount(const Slice& key, int32_t start_offset, int32_t end_offset,
                   int32_t* ret, bool have_range);
 
-  // Perform a bitwise operation between multiple keys (containing string values)
+  // Perform a bitwise operation between multiple keys
   // and store the result in the destination key
   Status BitOp(BitOpType op, const std::string& dest_key,
                const std::vector<std::string>& src_keys, int64_t* ret);
 
   // Return the position of the first bit set to 1 or 0 in a string
   // BitPos key 0
-  Status BitPos(const Slice& key, int32_t bit, int32_t* ret);
+  Status BitPos(const Slice& key, int32_t bit, int64_t* ret);
   // BitPos key 0 start
   Status BitPos(const Slice& key, int32_t bit,
-                int32_t start_offset, int32_t* ret);
+                int64_t start_offset, int64_t* ret);
   // BitPos key 0 start end
   Status BitPos(const Slice& key, int32_t bit,
-                int32_t start_offset, int32_t end_offset,
-                int32_t* ret);
-  
+                int64_t start_offset, int64_t end_offset,
+                int64_t* ret);
+
   // Decrements the number stored at key by decrement
   // return the value of key after the decrement
   Status Decrby(const Slice& key, int64_t value, int64_t* ret);
 
-  // Increments the number stored at key by increment. 
+  // Increments the number stored at key by increment.
   // If the key does not exist, it is set to 0 before performing the operation
   Status Incrby(const Slice& key, int64_t value, int64_t* ret);
